Backend Task Solution - Michal Cespiva
Hello Mews,

adding my two cents to the backend task solutions.

I am not sure whether this was expected from us but aside from implementing the GetExchangeRates method I took the liberty of changing a few things in the provided "scaffolding" code. Basically extending it here and there. However, the overall contract stays the same, i.e. the currencies array as the input and the list of exchange rates outputted to the console. Hope that's fine.

Thank you and have a great day!
Michal