Alberto Acuña - .Net challenge
This commit contains the implementation for the .Net challenge that we can find in this repository.

I tried to make the solution as straight forward as possible, having a `CZKExchangeRageApiProvider` that will take care of fetching the data of CZK currency. It inherits from `BaseExchangeRateApiProvider` so, if in the future we want to add other APIs, we all follow the same pattern. Additionally, this could help for future implementation of a cache or similar strategies in case it's a highly requested provider.