feat(t-shaped): Full stack task
Adds a full stack task as per Conor and our
T Shaped initiative.

Still gives autonomy to candidate for tech decisions,
but provides a starting point folks can leverage.

The idea is to have a backlog of 'real world' tasks
in a 'real world' project that simply is a product.

Ideally, down the road, each tribe may have its own
version of the project or backlog (or just a prioritized
item in the backlog) for tribe specific recs