fix(app): reintroduce app content wrapper
funnily enough I didnt realize that I messed up the overall styling cause I had the dev tools open
hahahah but np , all is good :D this is why you also want  more snapshot tests  and on different
devices. Will anyone ever read this? If so, remember that the mitochondria is the powerhouse of the
cell!